+++ src/components/custom-field-input-single-choice/custom-field-input-single-choice.jsx
@@ -7,102 +7,34 @@ import React, {
   useState,
 } from 'react';
 import PropTypes from 'prop-types';
-import AbstractCustomField from '../__internal__/abstract-custom-field/abstract-custom-field.jsx';
-import Icon from '../icon/icon.jsx';
-import IconButton from '../icon-button/icon-button.jsx';
-import iconClear from '../../svgs/clear.svg';
-import iconCaretDown from '../../svgs/caret-down.svg';
-import iconCaretDownDisabled from '../../svgs/caret-down-disabled.svg';
-import styles from './custom-field-input-single-choice.css';
-import Listbox from '../listbox/listbox.jsx';
+import Select from '../select/select.jsx';
 import ListOption from '../list-option/list-option.jsx';
-import NoOptions from '../no-options/no-options.jsx';
-import useValidation from '../../hooks/use-validation.jsx';
-import useDropdownClose from '../../hooks/use-dropdown-close.js';
 
 const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleChoice(props, ref) {
-  const [showOptions, setShowOptions] = useState(false);
   const [value, setValue] = useState(props.value);
-  const [searchValue, setSearchValue] = useState(undefined);
-  const inputRef = useRef();
   const backupRef = useRef();
   const selfRef = ref || backupRef;
 
-  const validationMessage = useValidation(props.readOnly, props.errorText, inputRef, false);
-  const refs = props.choices.map(() => createRef());
-  const caretIcon = (<Icon
-    className={styles['input-icon']}
-    icon={props.readOnly ? iconCaretDownDisabled : iconCaretDown}
-    label={props.readOnly ? 'Choice field is not editable' : 'Open choices listbox'}
-  />);
+  const listOptionRefs = props.choices.map(() => createRef());
 
-  const defaultValue = value ? value.label : '';
-
-  const wrapperRef = useRef(null);
-  const handleDropdownClose = () => {
-    setShowOptions(false);
-    setSearchValue(defaultValue);
-  };
-  useDropdownClose(wrapperRef, showOptions, handleDropdownClose);
-
-  const clear = () => {
-    setValue(undefined);
-    setSearchValue(undefined);
-    inputRef.current.focus();
-  };
-
-  const clearIcon = () => {
-    if (!props.readOnly && (value || searchValue)) {
-      return (<IconButton
-        className={styles['clear-button']}
-        icon={iconClear}
-        label={'Remove selected choice'}
-        onPress={clear}
-      />);
-    }
-
-    return undefined;
-  };
-
-  function onClick() {
-    if (!props.readOnly) setShowOptions(true);
-  }
-
-  function onKeyDown(event) {
-    switch (event.key) {
-      case 'Enter':
-        event.preventDefault();
-        if (!showOptions && !props.readOnly) setShowOptions(true);
-        break;
-      case 'Escape':
-        event.preventDefault();
-        setShowOptions(false);
-        break;
-      default:
+  function selectOnChangeHandler(event) {
+    if (event.target.value) {
+      setValue(props.choices.find(choice => choice.id === event.target.value.id));
+    } else {
+      setValue(undefined);
     }
   }
 
-  function getOptions() {
+  const listOptions = () => {
     const choices = {};
     props.choices.forEach((item, index) => { choices[item.id] = { ...item, index }; });
 
-    if (searchValue) {
-      const searchValueLowerCase = searchValue.toLowerCase();
-
-      return props.choices
-        .filter(item => item.label.toLowerCase().includes(searchValueLowerCase))
-        .map(item => choices[item.id]);
-    }
-
-    return props.choices.map(item => choices[item.id]);
-  }
-
-  const listOptions = (choices) => {
-    return choices
+    return props.choices
+      .map(item => choices[item.id])
       .map(item => (
         <ListOption
           key={item.id}
-          ref={refs[item.index]}
+          ref={listOptionRefs[item.index]}
           selected={value && item.id === value.id}
           value={{
             id: item.id,
@@ -113,26 +45,6 @@ const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleC
         </ListOption>
       ));
   };
-
-  function onSelectionChange(event) {
-    const selectedValue = event.target.value;
-    setValue(selectedValue);
-    setSearchValue(selectedValue.label);
-    setShowOptions(false);
-    inputRef.current.focus();
-  }
-
-  function onSearchChange(event) {
-    const newValue = event.target.value;
-
-    if (newValue === '') {
-      clear();
-      return;
-    }
-
-    setSearchValue(newValue);
-    setShowOptions(true);
-  }
 
   useImperativeHandle(selfRef, () => ({
     get dirty() {
@@ -150,48 +62,23 @@ const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleC
     props.onChange({ target: selfRef.current });
   }, [value]);
 
-  const choices = getOptions();
-
   return (
-    <div ref={wrapperRef} className={props.className}>
-      <AbstractCustomField
-        icon={caretIcon}
-        clear={clearIcon()}
-        id={props.id}
-        label={props.label}
-        name={props.name}
-        onChange={onSearchChange}
-        onClick={onClick}
-        onKeyDown={onKeyDown}
-        placeholder={props.placeholder}
-        readOnly={props.readOnly}
-        inputRef={inputRef}
-        required={props.required}
-        errorText={validationMessage}
-        value={searchValue || defaultValue}
-        inputRole={'combobox'}
-        ariaProps={{
-          autocomplete: 'list',
-          controls: `${props.id}-single-choice-listbox`,
-          expanded: showOptions,
-          haspopup: 'listbox',
-        }}
-      />
-      { showOptions && (
-        choices.length === 0 ? (<NoOptions className={styles['no-options']} />) : (
-          <Listbox
-            className={styles.dropdown}
-            id={`${props.id}-single-choice-listbox`}
-            labelledBy={`${props.id}-label`}
-            onChange={onSelectionChange}
-            refs={refs}
-            value={value}
-          >
-            { listOptions(choices) }
-          </Listbox>
-        )
-      )}
-    </div>
+    <Select
+      className={props.className}
+      displayValueEvaluator={selectValue => selectValue.label}
+      errorText={props.errorText}
+      id={props.id}
+      label={props.label}
+      listOptionRefs={listOptionRefs}
+      name={props.name}
+      onChange={selectOnChangeHandler}
+      placeholder={props.placeholder}
+      readOnly={props.readOnly}
+      required={props.required}
+      value={value}
+    >
+      { listOptions() }
+    </Select>
   );
 });
 
@@ -216,7 +103,7 @@ CustomFieldInputSingleChoice.propTypes = {
 
 CustomFieldInputSingleChoice.defaultProps = {
   choices: [],
-  className: styles.container,
+  className: undefined,
   onChange: () => {},
   placeholder: undefined,
   readOnly: false,