+++ CHANGELOG.md
@@ -8,6 +8,9 @@
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
   - Patch: Update PageHeader typography to match MDS specs
+  - Minor: Add new `Select` component
+  - Minor: Refactor `CustomFieldInputSingleChoice` to use `Select`
+  - Minor: Add new `setSelected` and `setVisible` methods to `ListOption` ref API
 </details>
 
 ## v0.43.3